+++ src/sempipe/cli/sem_file.py
@@ -203,6 +203,7 @@ _VERB_KEYS: Mapping[str, tuple[tuple[str, _KeySpec], ...]] = {
         ("schema-file", _str_key(_schema)),
         ("schema-from", _str_key(_flag("--schema-from"))),
         ("tally", _str_key(_flag("--tally"))),
+        ("explode", _str_key(_flag("--explode"))),
         ("fields", _list_key(_fields_arg)),
         *_COMMON_TAIL,
     ),